Add files via upload
The project is based on NATURAL LANGUAGE PROCESSING to classify the symptoms of coronavirus with common diseases like common flu, cough, body pains. Used vectorizer from the spacy library to vectorized the words and added them to the language pipeline. Used Support vector machine to classify the sentences from flu to COVID_19.